test(logic): split supplier feedstock tests under 400-line limit (#2847)
Extract shared fixtures and mineral-prospecting cases into separate files
so repo.logic_test_file_size passes on PR #3256.
